Move surface forward declarations to another file
